requirements.txt
flask
irene_api
bs4
